+++ Magic/CHANGELOG.md
@@ -35,6 +35,7 @@
  - Expose enchantment levels as attributes, currently only works on the mainhand item
  - Add axolotl and goat skins for shrink, airstrike and camera
  - Player currency balance now available as placeholder variables, e.g. $currency or @materia
+ - Brush selector now shows materia balance in inventory title
 
 ## Changes